pyproject.toml
[tool.poetry]
name = "vsporte-auth-server"
version = "0.1.0"
description = ""
authors = ["pivasikpivtsakov <elda250@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
alembic = "^1.13.0"
sqlalchemy = "^2.0.23"
pyjwt = "^2.8.0"
pydantic = "^2.5.2"
bcrypt = "^4.1.1"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
